Move ParseConsoleArguments out of Main.cpp
Facilitate future unit test calls on this function
